# backend/api/api.yaml
openapi: 3.0.3
info:
  title: MedMarket API
  version: 0.1.0
  description: Prescription price comparison platform API.

servers:
  - url: /

paths:
  /api/auth/register:
    post:
      operationId: registerUser
      summary: Register a new user account
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/auth/login:
    post:
      operationId: loginUser
      summary: Exchange credentials for a bearer token
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/auth/profile:
    get:
      operationId: getProfile
      summary: Get the authenticated user's profile
      tags: [auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/prescriptions/upload:
    post:
      operationId: uploadPrescription
      summary: Upload a prescription document and its metadata
      tags: [prescriptions]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [document, physicianName, medName, strengthValue, strengthUnit, quantity]
              properties:
                document:
                  type: string
                  format: binary
                  description: The prescription document (image or PDF)
                physicianName:
                  type: string
                medName:
                  type: string
                strengthValue:
                  type: string
                strengthUnit:
                  type: string
                quantity:
                  type: integer
      responses:
        '201':
          description: Prescription created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrescriptionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/prescriptions:
    get:
      operationId: listPrescriptions
      summary: List the authenticated user's prescriptions
      tags: [prescriptions]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The user's prescriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PrescriptionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/prescriptions/{id}/search:
    post:
      operationId: searchPrescriptionPrices
      summary: Search pharmacies for prices on a prescription's medication
      tags: [prescriptions]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The prescription to price
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Quotes from the pharmacies that stock the medication, cheapest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuoteResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    Address:
      type: object
      required: [street1, city, state, zip]
      properties:
        street1:
          type: string
        street2:
          type: string
        city:
          type: string
        state:
          type: string
          description: Two-letter state abbreviation
          minLength: 2
          maxLength: 2
        zip:
          type: string

    RegisterRequest:
      type: object
      required: [firstName, lastName, email, password]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        phone:
          type: string
        address:
          $ref: '#/components/schemas/Address'

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    TokenResponse:
      type: object
      required: [token]
      properties:
        token:
          type: string

    UserResponse:
      type: object
      required: [id, email, firstName, lastName]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        address:
          $ref: '#/components/schemas/Address'

    PrescriptionResponse:
      type: object
      required: [id, physicianName, medName, strengthValue, strengthUnit, quantity, documentUrl]
      properties:
        id:
          type: string
          format: uuid
        physicianName:
          type: string
        medName:
          type: string
        strengthValue:
          type: string
        strengthUnit:
          type: string
        quantity:
          type: integer
        documentUrl:
          type: string
          description: Short-lived presigned URL to the uploaded document

    QuoteResponse:
      type: object
      required: [pharmacyId, pharmacyName, pharmacyItemId, unitPriceCents, totalCents]
      properties:
        pharmacyId:
          type: string
          format: uuid
        pharmacyName:
          type: string
        pharmacyItemId:
          type: string
          description: The pharmacy's own identifier for the item (SKU, code, ...)
        unitPriceCents:
          type: integer
          format: int64
          description: Price per unit, in cents
        totalCents:
          type: integer
          format: int64
          description: Unit price multiplied by the prescription quantity, in cents

    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string